+++ bonsai-ai/bonsai_ai/_vendor/product_insights/_lib/pybond/__init__.py
@@ -1,7 +1,0 @@
-__author__ = 'Keegan Dahm'
-__version__ = '0.0.1'
-
-from . import encoders
-from . import types
-from .bond_field import BondField
-from .types import BondSchema