+++ bonsai-ai/bonsai_ai/_vendor/product_insights/_lib/pybond/types/bond_nullable.py
@@ -1,5 +1,0 @@
-from pybond.types import BondList
-
-
-class BondNullable(BondList):
-    pass