+++ bonsai-ai/bonsai_ai/_vendor/product_insights/log_event_drop_reason.py
@@ -1,8 +1,0 @@
-class LogEventDropReason(object):
-    """
-    Enum for reasons why an event was not added to the queue
-    """
-    EVENT_NOT_ADDED = -1
-    NO_EVENTS_ACCEPTED = -2
-    MAX_EVENTS_REACHED = -3
-    LOG_MANAGER_NOT_INITIALIZED = -4