+++ bonsai-ai/bonsai_ai/_vendor/product_insights/pipeline/__init__.py
@@ -1,5 +1,0 @@
-from __future__ import absolute_import
-
-from product_insights.pipeline.event_pipeline import EventPipeline
-
-__all__ = ['EventPipeline']